Tests: Move test fixtures to subfolder
