+++ client/state/reader/thumbnails/test/fixtures/sample-vimeo-response.js
